+++ ydb/public/sdk/cpp/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added support for the new inverted index type: JSON, intended to speed-up queries on Json or JsonDocument columns.
+
 * EXPERIMENTAL! Added `IProducer` interface to the SDK. This interface is used to write messages to a topic.
 Each message can be associated with a partitioning key, which is used to determine the partition to which the message will be written.